+++ src/Monitor/Monitor/ChangeLog.md
@@ -26,6 +26,20 @@
     * `New-AzDataCollectionEndpoint`
     * `Remove-AzDataCollectionEndpoint`
     * `Update-AzDataCollectionEndpoint`
+  * [Breaking Change] Action Group upgraded API version to stable 2023-01-01
+  * [Breaking Change] Use new and update cmdlets instead `Set-AzActionGroup` cmdlet
+  * The receiver used subtype cmdlets to create a replacement for command `New-AzActionGroupReceiver`
+    * New-AzActionGroupArmRoleReceiverObject
+    * New-AzActionGroupAutomationRunbookReceiverObject
+    * New-AzActionGroupAzureAppPushReceiverObject
+    * New-AzActionGroupAzureFunctionReceiverObject
+    * New-AzActionGroupEmailReceiverObject
+    * New-AzActionGroupEventHubReceiverObject
+    * New-AzActionGroupItsmReceiverObject
+    * New-AzActionGroupLogicAppReceiverObject
+    * New-AzActionGroupSmsReceiverObject
+    * New-AzActionGroupVoiceReceiverObject
+    * New-AzActionGroupWebhookReceiverObject
 
 ## Version 4.6.0
 * Fixed `Get-AzInsightsPrivateLinkScope` to support `ResourceId` parameter [#22568]